[IMP] new schema folder layout
