+++ .gitignore
@@ -1,3 +1,2 @@
 /target
 /.idea
-/flatbuffer_types/schema/sfu_generated.rs
