+++ flatbuffers/schema.fbs
@@ -1,4 +1,4 @@
-namespace sfu;
+namespace ws_api;
 
 table MediaSource {
     codec: string (required);
